--- entities/__init__.py
@@ -1,2 +1,0 @@
-import sympy
-x = sympy.symbols('x')
