--- utils/server_specials.py
@@ -1,7 +1,0 @@
-import uuid
-import datetime
-from config import Config
-
-
-def gen_token():
-    return uuid.uuid4().hex, datetime.datetime.utcnow() + datetime.timedelta(seconds=Config.TOKEN_LIFETIME_SEC)
